Make the InvalidSysctl exception more descriptive
The parameter being acted on by sysctl() call is determined by an arbitrarily
sized array of integers called MIB. The current version of the code only
prints the first element of the MIB, which isn't descriptive at all. This
change tries to print 4 elements, which should be enough for most of sysctls
in the wild.

While there, make type signatures of low-level functions match their C versions.

The problem was raised by a FreeBSD user:
https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=289022