Fix issues with interpreted EH
The exception handling with interpreter enabled was not working properly
when the managed code exception handling code was interpreted too. There
were couple of places where we try to unwind to the first managed frame
using native unwind that obviously doesn't work for fully interpreted
code.
There is also an assert in the ExceptionHandling.cs that checks sanity
of the handling frame SP, but it assumes that managed locals are stored on
the processor stack. This is not the case for interpreted code.

This change fixes those places, namely:
* Updating floating point registers for Frames. It now unwinds upto a
  specific SP instead of searching for managed frame except for
  InlinedCallFrame case when that frame is in JIT/AOT code.
* Resuming after catch is fixed and also simplified a lot.
* Disable the sanity assert when runtime is compiled with interpreter. I
  don't think there is much value in that assert.

It also enables two pinvoke tests that were fixed in the past already,
but not enabled yet.