[mono] Fix g_str_hash skipping the first byte and remove mono_metadata_str_hash (#119208)
Port of https://github.com/mono/mono/issues/21844

`g_str_hash` was skipping the first byte because the `*p++` increment happens before the hash calculation, so when `*p` is referenced inside the loop, it's actually looking at the character after the one that was just tested.

We can also remove `mono_metadata_str_hash` which has similar logic (but isn't technically affected since it was already accounting for the first byte in  initializating `guint hash = *p`) since we no longer support linking to a different glib - we always link to eglib - and so the original reason for having a separate function no longer applies.