Limit doublemapper maximum size
The double mapping currently sets the maximum size of the double mapped memory to 2TB.
However, that causes the runtime to fail to start if there is a rlimit for max file
size set to a smaller value.

This change fixed the problem by clipping the max double mapped memory size by that
rlimit and also uses much more conservative size even without the file size limit set.
It is now set to the total amount of physical memory by default and also clipped by
the virtual memory rlimit.

Close #117819