Use a SafeHandle when duplicating a certificate context.
CertDuplicateCertificateContext does not ensure the CERT_CONTEXT pointer it is incrementing has not been freed.
If the duplicate and dispose race, duplicating a disposed handle will lead to unspecified behavior.

For .NET 10, we can use the SafeHandle around the certificate before we duplicate the handle. For OOB, we don't have
access to the internals, so we will continue to use the IntPtr Handle property.