[host] Set SWIFT_BACKTRACE to enable=no on OSX
For reasons yet to be confirmed, when an application throws an unhandled exception, the swift backtrace handler appears to be on for various versions of MacOS that we support.

This leads to a messy and unclear output of a backtrace on the main host process. It can be worked around by setting the env variable SWIFT_BACKTRACE=enable=no.

Contributes to https://github.com/dotnet/runtime/issues/118823