[nativeaot] Don't overwrite `$(LinkNativeDependsOn)` property
The `$(LinkNativeDependsOn)` property is used by the `LinkNative`
target to list its dependency targets. Unfortunately, its value
is unconditionally overwritten in `Microsoft.NETCore.Native.targets`,
making it impossible for 3rd parties to add custom dependency targets
to the `LinkNative` target.

Change the code to make sure that previous value of the property
is preserved.