Replace expansion of MINT_ICALLSIG to cookies with signature encoding
Previously was "encoding" into a cookie but in do_icall, which relied on having a non-null MonoMethodSignature, which is not always the case.
Instead we expand MintIcallSig with all permutations of arg types (generated from an external app) and encode directly to the enum from the MonoMethodSignature (via interp_get_icall_sig) .
A decode method provided isn't required here but might be useful for testing methods.