Remove `SR.Argument_NeedStructWithNoRefs`
This change removes the `SR.Argument_NeedStructWithNoRefs` resource string and updates the corresponding `ArgumentException` message to be more descriptive.

The old message was:

```
The specified Type must be a struct containing no references.
```

This has been replaced with:

```
Cannot use type '{0}'. Only value types without pointers or references are supported.
```