JIT: Avoid inverting already-inverted loops that are IV tested
Graph-based loop inversion works by inverting the loop such that we do
not jump into the loop to immediately test a condition. However, we are
not detecting the case where the loop is already inverted because its
latch is an exit that tests a different condition. If the loop has
multiple exits that means we may end up creating a suboptimal latch
condition.

Detect this case based on the pattern-based IV analysis that we already
have.