[interp] Treat 'unsigned' fp->int conversions like signed ones (#119549)
We have some test cases that use .un conversion opcodes to convert floats to ints. The spec doesn't say what this does, and it doesn't make much sense. What the JIT does is ignore the .un modifier in this scenario, so this PR updates the interpreter to do the same thing.

This is documented in the augments (https://github.com/dotnet/runtime/blob/main/docs/design/specs/Ecma-335-Augments.md#unsigned-data-conversion-with-overflow-detection)