Allow using UTF-8 in CompilerGeneratedNames
The managed type system switched to UTF-8, The `CompilerGeneratedNames` class is forcing us to materialize hundreds of thousands of UTF-16 strings (in a `dotnet new webapiaot` sample). Add overloads that take UTF-8.

I was thinking to do some code reuse with clever use of structs and interfaces but this is one of the cases where simpler is just better, so a copy it is.