ILLink.RoslynAnalyzer: fix hole for new constraint (#119291)
The new constraint processing was happening in a separate code path
from the other generic argument processing, and that code path didn't
see the implicit call to the base ctor (because it didn't use the
IOperation tree). This fixes it by moving the new constraint
processing into `GenericArgumentDataFlow`.

Unlike ILC, the analyzer can't just treat `new()` the same as
`PublicParameterlessConstructors`, because it needs to produce
analysis warnings for `RequiresDynamicCode` or `RequiresAssemblyFiles`
accessed via `new()` constraint, even when trim analysis is disabled
(the annotation is treated as reflection access, but `new()` is not).