Add a flag to hostfxr_resolve_sdk2_flags_t to silence error messages (#119729)
hostfxr_resolve_sdk2 will print error messages to stderr by default. We can add a flag to the enum to allow for silencing these errors when we want to try to resolve an sdk, but a failure to resolve shouldn't be surfaced to the console.

See https://github.com/dotnet/sdk/issues/50632 for context.