Fix CodeQL detecting EVP_chacha20_poly1305 during cmake configure (#119887)
Some internal CodeQL rules are picking up this instance but since CMake generates a random directory path for the function detection it results in a lot of duplicate bugs.

Use the CodeQL comment syntax to suppress it. We need to use check_source_compiles() instead of check_function_exists() so that we can inject the suppression comment.