JIT: enable instrumentation for inlinees (#119658)
If we are doing an optimized+instrumented jit pass (like we do for R2R methods)
allow the inlinees to be instrumented. This gives us a chance to collect profile
data for methods that are always inlined.

All inlinees currently share the same profile data segment with each other and
with a root compilation of the method (if any). So this profile is "context-free".

If there is a schema mismatch (say there is a stale pre-existing R2R schema) then
subsequent instrumentation will fail. This is something we need to keep an eye on.
Right now we can't distinguish this failure from other kinds of schema allocation
failures.

Closes #44372. Closes #91938.