Map SIMD types directly from class names
The current implementation derives a size based on the identified SIMD
type, and then uses the size to derive the node type. It should instead
directly derive the node type from the identified SIMD type, because
some SIMD types will not have a statically known size, and this size may
conflict with other SIMD types.