Simplify ComInterfaceGenerator's Initialize() method (#119360)
The ComInterfaceGenerator takes a significant amount of time in its Initialize() method setting up an elaborate pipeline, even for compilations that don't use generated COM interfaces. In particular, it spent a lot of time in the series of Zip that preceded writing the source out. This PR reduces the number of incremental steps the generator creates, reducing the Initialize() time for projects that don't need the generator. This will likely reduce incremental performance slightly, but the tradeoff seems worth it considering the vast majority of projects won't need this generator.

The change moves all steps in the pipeline preceding the creation of the IncrementalMethodStubGenerationContext to a single step. It also moves the generation of all syntaxes into a single step rather than creating them in parallel steps and Ziping them together. The ItemAndSyntaxes type is used to keep the InterfaceContext (which is required in the source-writing step) in the collection so it doesn't need to be re-zipped in before writing out.

After this change, there are a number of other simplifications we can make (we don't need so many intermediate steps within the first step, the DiagnosticOr concept can be simplified) that I'd like to add in follow-up PRs.

Addresses #119183