Add warning for annotated extension property (#120005)
Fixes https://github.com/dotnet/runtime/issues/119113 by adding a new
warning that is shown by the Roslyn analyzer when attempting to
annotate an extension property with
DynamicallyAccessedMembersAttribute.

The warning is also produced by ILLink and ILC, but only when the
compiler-generated extension metadata type (which contains the
property in IL) is referenced, for example via reflection or when the
assembly is rooted. So in practice this will be caught just by the
analyzer.

See comments in the code that explain some nuances of the test
approach (we need a way to assert `ExpectedWarning`s on the extension
properties, but not the compiler-generated methods in the extension
metadata type).

Testing this uncovered a test bug which was hiding
https://github.com/dotnet/runtime/issues/120004.