Fix ResourceUpdater.Dispose() (#120029)
The apphost ResourceUpdater doesn't dispose of the PEReader when leaveOpen is true, but the leaveOpen should only apply to the FileStream passed to the constructor. The PEReader is not accessible from the creator and cannot be disposed of elsewhere. This causes some race conditions, especially in cases where the PEReader ends up mapping the file into memory.

Fixes dotnet/sdk#50784