WASM enable GC support (#120175)
* Disable PerfTracing on WASM. It was disabled in one spot but not all.

Disable more finalizer logic in WASM.

Check for null during GC trigger routine.

Fix check in stresslog logic by defining a non-zero thread ID on WASM.

* Refactor for consistency and modern C++ style.

* Fix GcInfo for WASM - missed a bit when decoding GcInfoHeaderFlags.

Update the stack limit on WASM. It can be zero, but CoreCLR assumes
this is a failure. As a mitigation we simply subtract a pointer size from
the stack limit, thus reducing the stack by a pointer but making it non-zero.

* Fix WASM release build.

* Update WASM logic for thread ID.

* Ensure the entire value of tid is set.

* Feedback

* Feedback