Move hosting tests to XUnit 3
The primary advantage in this change is that test assemblies are now
executables, not libraries. This makes debugging easier, as the
executable can be run directly under the debugger. It also makes
deployment of the tests easier, as we only need to copy the executable
and set DOTNET_ROOT to the pre-built framework.

This is also a kind of trial-balloon. Many other tests (esp. libraries
tests) in this repo would benefit from the execution advantages.
However, they are built in more configurations today and are more
complex. The host tests are an easier lift.