Fix CopyOSContext on amd64 (#120300)
* Fix CopyOSContext on amd64

The CopyOSContext was skipping floating point registers on amd64. There
was an obsolete comment that on some older Windows, the vector registers
are not present in the context. I've found to be a problem when porting
nativeAOT EH to CoreCLR and to make the context flags correct, I've
removed the CONTEXT_FLOATING_POINT from the ContextFlags. But it now
turned out that it prevents EH from correctly restoring non-volatile
floating point registers after a catch. Since he ContextFlags don't
contain the CONTEXT_FLOATING_POINT, the floating point registers were
not restored during the resume.

This change fixes it by removing the amd64 specific handling and always
copying the whole CONTEXT structure. It just masks off the
CONTEXT_XSTATE, as the XSTATE is stored out of the basic CONTEXT.

Close #118564

* Fix build break - CONTEXT_XSTATE exists on some targets only