Arm64-SVE: Fixes for returning Vector<T> through the return buffer
Invalidates debug information for return buffer when it holds TYP_SIMD. Computing
an accurate stack offset is not possible in AOT, so we avoid introducing this support.
Better information can be provided when the return value is passed in a vector register.

Prevent the hidden return buffer optimization when the local has TYP_SIMD. This
optimization will no longer apply when the full vector calling convention is supported,
so we avoid building support now.