Pass Vector<T> by reference when InstructionSet_VectorT is available
Prevents Vector<T> being treated as a 128-bit HFA when the new SVE instruction
set for Vector<T> is available. A new HFA type for SVE-backed Vector<T> will
be implemented in place of this.

Also take some time to improve the HFA recognition code. HFA type can be
decided immediately from the class name for fixed size types. Create a central
function for this decision.