Implement UnknownSizeFrame for locals with unknown size
Implements a simple bump allocator for TYP_SIMD and TYP_MASK. Locals are
allocated to this space when lvaIsUnknownSizeLocal is true for the variable.

The frame is implemented on ARM64 as two homogenenous blocks containing either
TYP_SIMD or TYP_MASK locals. The x19 register is reserved for addressing locals
in the block. Updates codegen for SVE memory transfer instructions to accept
indices in multiples of the vector length (or VL / 8 for masks) instead of
deriving them from the size of the local.