Fix GCFrame allocated by interpreted code handling
There are few cases where GCFrame is created by managed code as a
managed local. When that code is interpreted, this local is on the
interpreter stack which is NOT the processor stack. So various debug
checks and also removing these GCFrames from the per-thread chain
doesn't work.

This change adds an "os stack location" member to the GCFrame that
is set to "this" for GCFrames allocated in native code. For the ones
allocated by the interpreted managed code, it is set to the
InterpMethodContextFrame of the frame that created it. That value is
then used instead of the raw GCFrame pointer to compare locations of
these frames.