[mono][sgen] Enable configuration of nursery allowance ratio
default-allowance-ratio represents the minimum amount of heap growth that we allow before triggering a new major collection, as a relative multiplier to the nursery size. Setting this option was always skipped if max_heap is not set, which is the common case (max_heap requires specific gc knobs to be configured). This option has nothing to do with the max_heap so make sure it is correctly initialized.