Reduce constraint loading by not forcing constraints to be loaded eagerly - Reduce the set of constraints that need to be loaded for Bounds and cast checking
TODO:
There is a path in InitTypeContext which I have #ifdef'd out as I don't understand the comment
The accessibility checking needs full loading now, as we don't have a scheme to just look at all the typedefs in a signature, and instead need to do a load of the full type and work from there. This could be revisited.