Fix ARM64 ThreadAbort issue
There is a problem with RtlRestoreContext not restoring context during
ThreadAbort if that abort comes from injected APC callback on ARM64 and
the processor supports SVE. In that case, the context provided by the
APC callback can contain XSTATE and when we start walking stack from
that context, we copy that context into the REGDISPLAY. The problem is
that we copy the ContextFlags without changes, so if they contained
CONTEXT_XSTATE flag, it is kept set even though the REGDISPLAY has only
plain old CONTEXT without any xstate.
When we call RtlRestoreContext in the ResumeAfterCatch, it fails because
the XSTATE is not valid and so it returns. That's unexpected and we end
up crashing with an assert.

The fix clears the CONTEXT_XSTATE in ResumableFrame::UpdateRegDisplay so
that the REGDISPLAY ContextFlags are validly representing the context.

Close #120437