Bring back the reader-writer lock for RcwCache
Instead of writing to `_cache` while under a reader lock, delay the cleanup until we can acquire a writer lock.

We do this instead of doing an upgradeable lock as `ReaderWriterLockSlim` only supports one upgradeable reader thread at a time, removing the benefit of having an rwlock in the first place.