Fix interpreter compiler branch elimination (#120768)
The code emission was eliminating branches at the end of a basic block
going to the next basic block. This should only happen in case when both
blocks are on the same overlapping clause depth though. It was breaking
couple of EH tests where it let execution flow from the end of a try
block to an unrelated block.