[clr-interp] Add a scheme for adding peephole optimizations for known sequences of IL opcodes - This is needed to ensure that some of our SIMD tests finish in a vaguely reasonable timeframe - The peeps implmented here make that almost possible, although the behavior in CI hasn't yet been verified - Peeps implemented   - stloc/ldloc  ... This is a minor improvement, but will be needed to handle a future optimization around allowing the il stack to have constant values   - box/unbox.any - Removes unnecessary boxing/unboxing   - typeof(T)==typeof(Y) - This allows us to handle the type testing specialization behavior that is used heavily in the BCL   - typeof(T).IsValueType - Used in the Unsafe.BitCast function
Not yet implemented is giving the interpreter stack a concept of constant values, so that we can optimize brtrue/brfalse and friends into unconditional branches. With this set of changes the if (typeof(T)==typeof(int)) { ... } else if (typeof(T)==typeof(float)) pattern is *much* faster than before, but the not taken paths are still fully generated interpreter code, and there are still many branches.