Fix issue with return / goto from catch inside catch that exits the outer catch (#120821)
When there is a try/catch inside a catch handler and there is a return
in the inner catch or goto going out of the outer catch handler, we were
resuming after the inner catch at the end of the method / goto target,
so we were not returning from the call to the outer catch.
We need to inject a catch leave island that the inner catch returns as
its resume after catch location and that island can then return from the
outer catch.
This mechanism needs to be part of the finally call islands chain so
that
returns from catches / calls to finallys happen correctly and in the
right order.
This change extends the finally call islands chain to implement that.

This change also fixes a problem with ordering of emission of the code
for these islands and the basic blocks linking in the `GenerateCode`
method.
It also fixes a bug in `GetNativeRangeForClause` that was in some edge
cases getting incorrectly shorter range.

To make reviewing a bit easier, I've split this PR into two commits. The
first adds the actual functionality and the second just renames things
to match the fact that the islands are no longer finally call islands
only.