[RISC-V] Move comparison transformations from codegen to lowering (#118270)
Codegen was not a good place for these transformations, it resulted in
contorted logic and complicated register allocation. Every type of
comparison benefits: branch, floating, and integer but that was
especially true for non-branch integer comparisons -- only less-than is
available so any other comparison must be achieved with additional
operations which are much easier to insert/transform to at node level.

Inserting sign-extensions (and recognizing when it is necessary) for
comparands is also easier at node level. This PR introduces only simple
local checks, more comprehensive sign-extension elimination will be
attempted in subsequent PRs.

Part of #84834, cc @dotnet/samsung