[clr-interp] Fix additional iteration in Copy_Ref implementation on arm64 (#120879)
After a loop iteration is done, the code is subtracting 16, obtaining
the current remaining counter. If the remaining counter is bigger than 0
then we do another iteration. This is wrong, because we can only do
another 16 byte copy iteration if the remaining counter is bigger or
equal to 16. In order to be correct, the code would have to branch to
the very beginning of the macro. In case of negative result, we would
fallback to 8 byte copy, but, if the result is negative, it means we
already overcopied, so this implementation is wrong.

The fix makes it such that when we subtract 16 after an iteration, we
don't obtain the current remaining counter, rather the potential
remaining counter, in case we did another copy iteration. If this
counter is positive, then it is safe to do another iteration, if it is
negative then we can't copy 16 bytes at a time, fallback to the next
case.

The PR also removes redundant looping for 8 bytes copy scenario, which
can be hit at most once.