Fix ARM64 interpreter asm helpers (#120958)
Running the CoreCLR tests fully interpreted on macOS arm64 has revealed
a number of bugs in the assembler helper routines for call stubs.
* Load_Stack on macOS arm64 can get size that's not a multiple of 8
* InterpreterStubRetBuff was loading the return buffer address from a
slot that is usually not initialized by the x8. And the x8 is actually
valid at that place, so we can get the value from it.
* The Copy_Ref was fixed in another PR recently, but the same bug that
was fixed for the copying by 16 bytes was there for copying by 8 and 1
byte.
* The CallJittedMethodRetXXX functions were restoring the x2 from a SP
relative address, but SP was modified after the x2 was stored there. The
FP relative offset is stable, so I've switched to that one.

This change fixes all of these.