JIT: Add an opportunistic optimization to reuse continuation instances
This adds an optimization that tries to reuse the continuation instance
that was used to resume a runtime async function if it suspends again.

This optimization is not reliable. It only kicks in if the same
suspension point resumes and then suspends again. However, it is useful
to get an idea of how costly the continuation allocation part is when
looking at the gap between async1 and runtime async in raw
suspension/resumption performance.

The optimization is controlled by
`DOTNET_JitOpportunisticContinuationReuse` and is disabled by default.

Despite the above this is not a very costly transformation, so we may
consider enabling it even if it does not kick in very reliably.