Packing Legacy cDAC apis (#121081)
This PR does some work and refactoring in order to facilitate a private
version of the cDAC.

1. Assuming that we may want to access some of the legacy APIs from the
private cDAC, I have pulled the legacy code into a separate package and
made the COM interop classes public for use.
2. I have added a mechanism to add new contract implementations to a
target upon instantiation of the target. This way the private cDAC can
add its own contracts.
3. Exposing GetContract<TContract>. The private cDAC can in the future
call GetContract with contracts not found in the default contract
registry. This will probably be hidden behind an extension method but we
need GetContract public for implementation.

---------

Co-authored-by: Jeremy Koritzinsky <jkoritzinsky@gmail.com>