Handle emscripten define of "unix" during compilation
The Emscripten tool chain defines 'unix' during compilation.
This meant that during the data contract generation the "unix"
string was converted to "1" and interpreted as "windows".
See `./src/mono/browser/emsdk/emscripten/emcc -dM -E - < /dev/null | grep unix`.

The fix here is to create a specific OS for the browser and
have that run first. Place some defensive checks in place
for the other OS names that aren't all that unique either.